tap event, swipe event fixes:
tap events fired a double-tap unless you prevented default on touchend for the document. Which ended up breaking webpages. So now touchme passively listens to events and fires accordingly.

swipe events fired all the time because they didn't check the deltas to see if the user actually moved in the first place.